Move loading landXmlPath to separate hook because it's used in deviations as well now
