+++ src/features/followPath/routes/pathList.tsx
@@ -22,10 +22,9 @@ import { highlightActions, useDispatchHighlighted, useHighlighted } from "contex
 import { measureActions, singleCylinderOptions } from "features/measure";
 import { Picker, renderActions, selectPicker } from "features/render/renderSlice";
 import { AsyncStatus, hasFinished } from "types/misc";
-import { getObjectNameFromPath, getParentPath } from "utils/objectData";
-import { searchByPatterns } from "utils/search";
 
 import { followPathActions, selectFollowCylindersFrom, selectLandXmlPaths } from "../followPathSlice";
+import { useLoadLandXmlPath } from "../hooks/useLoadLandXmlPath";
 import { useFollowPathFromIds } from "../useFollowPathFromIds";
 import { usePathMeasureObjects } from "../usePathMeasureObjects";
 
@@ -33,7 +32,7 @@ export function PathList() {
     const theme = useTheme();
     const history = useHistory<{ prevPath?: string }>();
     const {
-        state: { db, view },
+        state: { view },
     } = useExplorerGlobals(true);
     const highlighted = useHighlighted().idArr;
     const dispatchHighlighted = useDispatchHighlighted();
@@ -62,54 +61,7 @@ export function PathList() {
         }
     }, [selectingPos, highlighted, dispatch]);
 
-    useEffect(() => {
-        if (landXmlPaths.status === AsyncStatus.Initial) {
-            getLandXmlPaths();
-        }
-
-        async function getLandXmlPaths() {
-            dispatch(followPathActions.setPaths({ status: AsyncStatus.Loading }));
-
-            try {
-                let paths = [] as {
-                    id: number;
-                    name: string;
-                }[];
-
-                await searchByPatterns({
-                    db,
-                    searchPatterns: [{ property: "Novorender/PathId" }],
-                    callback: (refs) =>
-                        (paths = paths.concat(
-                            refs.map(({ path, id }) => ({ id, name: getObjectNameFromPath(getParentPath(path)) }))
-                        )),
-                });
-
-                if (paths.length == 0) {
-                    //Legacy
-                    await searchByPatterns({
-                        db,
-                        searchPatterns: [{ property: "Novorender/Path", value: "true", exact: true }],
-                        callback: (refs) =>
-                            (paths = paths.concat(
-                                refs.map(({ path, id }) => ({ id, name: getObjectNameFromPath(getParentPath(path)) }))
-                            )),
-                    });
-                }
-
-                paths.sort((a, b) => a.name.localeCompare(b.name, "en", { sensitivity: "accent" }));
-                dispatch(followPathActions.setPaths({ status: AsyncStatus.Success, data: paths }));
-            } catch (e) {
-                console.warn(e);
-                dispatch(
-                    followPathActions.setPaths({
-                        status: AsyncStatus.Error,
-                        msg: "Failed to load list of paths to follow.",
-                    })
-                );
-            }
-        }
-    }, [db, landXmlPaths, dispatch]);
+    useLoadLandXmlPath();
 
     const isLoading =
         !hasFinished(landXmlPaths) ||
